Fix typos with test_client and test_server
